data_processing_pipeline/requirements.txt
# jupyterlab==4.3.5
# numpy==2.2.2
# pandas==2.2.3
# matplotlib==3.10.0
# seaborn==0.13.2
# scipy==1.15.1
# scikit-learn==1.6.1
# pyspark==3.5.5
pandas==1.5.3
pyspark==3.4.0
jupyterlab==3.6.3
matplotlib==3.7.1
numpy==1.24.3
python-dateutil==2.8.2